Fix `$ref` parameters not being dereferenced
`$ref` parameters are often used in the case of shared GET parameters. This
is so we don't have to redefine common parameters like `page[number]`. For
example:

    "paths": {
      "/messaging_profiles": {
	"get": {
	  "summary": "List all messaging profiles",
	  "parameters": [
	    {
	      "$ref": "#/components/parameters/PageNumber"
	    },

	    ...
	  ]
        }
      }
    }

These weren't taken into account when writing the spec parser. When
attempting to use one of these parameters, you would receive this error:

    {
      "error": {
        "message": "Request validation error: validator 0xc0003f81b0 failed: additional properties are not allowed: page[number]",
        "type": "invalid_request_error"
      }
    }

This commit fixes the parser to dereference these `$ref` parameters.