When no response `type` is provided, assume an `object`
An update to our spec has removed the hardcoded definition of the object
being `"type": "object"`, which appears to be valid Swagger/OpenAPI.o

This commit updates the service to no longer expect a `type` to be
provided. If no `type` is provided but the `properties` are, we (I
believe, safely) assume the type is an `object`.