Allow inline responses. Don't assume `#/components/...` $ref
Our previous spec heavily utilised the `components` section
(https://swagger.io/docs/specification/components/) to define responses
and this service was written to expect that behavior (and only that
behavior).

We've since updated our spec to now inline all responses, which broke
this service. This commit adds a check to account for whether the
response is inlined or a referenced component.