PORTAL-1014: Generate response examples from attribute level examples
One noticable side affect in this commit is that
`generateSyntheticFixture` no longer pays attention to required
parameters. Instead, it generates a fixture based on all parameters.

This is to allow more leniency in the provided spec file. It's possible
(and currently true, at the time of this commit) that required
parameters aren't listed as required.