PORTAL-980: Fetch spec from Github on startup
Currently, we bundle the API spec (spec3.json) within the application as
a go-bindata asset. This means, when the API spec gets updated, we need
to release a new version of the service. New Github release, homebrew
upload, docker image, etc.

This update changes the service to fetch the live spec file from Github
on startup. Caching it to speed up future runs.

The spec file continue to exists within the bundled application for the
sake of unit tests.